// apps/accounts-service/src/core/types/db.ts
import type { ColumnType } from "kysely";

export type AccountType =
  | "ASSET"
  | "BANK"
  | "CASH"
  | "CREDIT"
  | "INVESTMENT"
  | "LIABILITY";

export type Generated<T> =
  T extends ColumnType<infer S, infer I, infer U>
    ? ColumnType<S, I | undefined, U>
    : ColumnType<T, T | undefined, T>;

export type Numeric = ColumnType<number, number | string, number | string>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Accounts {
  accountType: AccountType;
  color: string | null;
  createdAt: Generated<Timestamp>;
  currency: Generated<string>;
  description: string | null;
  icon: string | null;
  id: string;
  initialBalance: Generated<Numeric>;
  isActive: Generated<boolean>;
  lastReconciled: Timestamp | null;
  name: string;
  updatedAt: Generated<Timestamp>;
  userId: string;
}

export interface DB {
  accounts: Accounts;
}